requirements.txt
# =============================================================================
# Requirements for the Multi-Layered, Multilingual Text Analysis Engine
#
# To install all dependencies into your virtual environment, run:
# pip install -r requirements.txt
# =============================================================================

# --- Core Deep Learning Framework ---
# PyTorch is the foundation for our model training and inference.
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# --- Hugging Face Ecosystem ---
# For loading transformer models (XLM-R), tokenizers, and optimizers.
transformers==4.41.2
# For training optimizations like device mapping and mixed precision.
accelerate==0.30.1
# For model quantization (used during data generation).
bitsandbytes==0.43.1

# --- Data Science & Machine Learning ---
# For data manipulation (reading CSVs).
pandas==2.2.2
# Core numerical library, a dependency for most ML packages.
numpy==1.26.4
# For label encoding and performance metrics (classification_report, confusion_matrix).
scikit-learn==1.4.2

# --- Web API & Serving (for the FastAPI server) ---
# The core web framework.
fastapi==0.111.0
# The high-performance web server to run FastAPI.
uvicorn[standard]==0.29.0
# Required for handling file uploads (for the OCR endpoint).
python-multipart==0.0.9

# --- Data Processing & Utilities ---
# For converting emojis to text during preprocessing.
emoji==2.11.1
# For creating progress bars in our scripts.
tqdm==4.66.4
# The Python wrapper for the Tesseract OCR engine.
pytesseract==0.3.10
# For image processing, required by pytesseract.
Pillow==10.3.0

# --- Visualization ---
# For plotting charts and confusion matrices in the evaluation script.
matplotlib==3.8.4
# For creating prettier heatmaps (confusion matrices).
seaborn==0.13.2